backend/requirements.txt
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.30.0
layoutparser>=0.3.4
pytesseract>=0.3.10
opencv-python>=4.7.0.72
Pillow>=9.4.0
pdf2image>=1.16.2
spacy>=3.5.0
python-dotenv>=0.21.1
pandas>=1.5.3
numpy>=1.24.2
scikit-learn>=1.2.2
PyPDF2>=3.0.1
reportlab>=3.6.12
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
streamlit>=1.22.0
ultralytics>=8.0.0